requirements.txt
Django==1.8.4
django-user-accounts
django-bootstrap-form
django-taggit
django-taggit-templatetags
django-notifications-hq
django-user-sessions
redis
djangorestframework
Pillow
django-model-utils
django-ipware